refactor(cloudsmith): split package operations and extract common code
