+++ pkg/integrations/cloudsmith/common.go
@@ -10,6 +10,7 @@ import (
 	"github.com/mitchellh/mapstructure"
 	"github.com/superplanehq/superplane/pkg/configuration"
 	"github.com/superplanehq/superplane/pkg/core"
+	"github.com/superplanehq/superplane/test/support/contexts"
 )
 
 const (
@@ -121,4 +122,32 @@ func defaultProcessQueueItem(ctx core.ProcessQueueContext) (*uuid.UUID, error) {
 
 func defaultHandleWebhook(ctx core.WebhookRequestContext) (int, *core.WebhookResponseBody, error) {
 	return http.StatusOK, nil, nil
+}
+
+func cloudsmithPackageExecutionContext(httpContext *contexts.HTTPContext, executionState *contexts.ExecutionStateContext) core.ExecutionContext {
+	return cloudsmithPackageExecutionContextWithConfiguration(
+		httpContext,
+		executionState,
+		map[string]any{
+			"repository": "acme/production",
+			"package":    "pkg_123",
+		},
+	)
+}
+
+func cloudsmithPackageExecutionContextWithConfiguration(
+	httpContext *contexts.HTTPContext,
+	executionState *contexts.ExecutionStateContext,
+	configuration map[string]any,
+) core.ExecutionContext {
+	return core.ExecutionContext{
+		Configuration: configuration,
+		HTTP:          httpContext,
+		Integration: &contexts.IntegrationContext{
+			Configuration: map[string]any{
+				"apiKey": "test-key",
+			},
+		},
+		ExecutionState: executionState,
+	}
 }
